Added a "lerp" and "once" prop to Favicon
Added a lerp and once option to favicon, lerp will cause the animation to play in reverse on completion and once, when animated is false and once is true, will cause the animation to play once. Once and lerp also work together in that when once and lerp are both true, the favicon will play the reverse animation to return to the original state. Otherwise it will simply return to the initial position.